First steps for validation function
